+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 * Add rounding postings only when required ([issue #9](https://github.com/beancount/beancount2ledger/issues/9))
 * Avoid printing too much precision for a currency ([issue #21](https://github.com/beancount/beancount2ledger/issues/21))
+* Avoid creating two or more postings with null amount ([issue #23](https://github.com/beancount/beancount2ledger/issues/23))
 * Add config option `indent`
 * Show metadata with hledger output
 * Support setting auxiliary dates and posting dates from metadata ([issue #14](https://github.com/beancount/beancount2ledger/issues/14))